requirements.txt
kaggle==1.5.5
python-dotenv==0.10.3
pytz==2019.2
tweepy==3.8.0
matplotlib==3.1.1
